requirements.txt
alembic==0.9.9
certifi==2018.4.16
click==6.7
cloudinary==1.12.0
Flask==1.0.2
Flask-Migrate==1.8.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.1
funcsigs==1.0.2
gunicorn==19.4.5
imgurpython==1.1.7
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
mock==2.0.0
MySQL-python==1.2.5
pbr==4.0.4
Pillow==5.1.0
psycopg2==2.7.5
python-dateutil==2.7.3
python-dotenv==0.8.2
python-editor==1.0.3
requests==2.9.1
six==1.11.0
SQLAlchemy==1.2.8
urllib3==1.23
Werkzeug==0.14.1
